Alterando as URLs para o novo servidor, alterando o modelo Onibus para comportar mais informaçoes

+++ Busao/Ponto.h
@@ -12,7 +12,7 @@
 
 @interface Ponto : NSObject <MKAnnotation>
 
-@property(nonatomic, readonly) long identificador;
+@property(nonatomic, readonly) int identificador;
 @property(nonatomic, readonly) NSString *nome;
 @property(nonatomic, readonly) NSString *descricao;
 @property(nonatomic, readonly) NSArray *onibuses;
